[ui][memory_saver] Fix MemorySaverResourceView not to overflow
MemorySaverResourceView was previously instantiated with a plain "int"
argument, which would overflow if more than 2GB of memory had been
"saved".

Bug: 326845945
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5317681
Auto-Submit: Wez <wez@chromium.org>
Commit-Queue: Alison Gale <agale@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1265240}

+++ chrome/browser/ui/views/performance_controls/memory_saver_resource_view.h
@@ -14,7 +14,7 @@ FASTVC-BENCH-CORPUS:chromium-main-v1-943b94ae-1c74-4335-94fa-ceb4d277cea8
 // visualization to contextualize those savings.
 class MemorySaverResourceView : public views::FlexLayoutView {
  public:
-  explicit MemorySaverResourceView(const int memory_savings_bytes);
+  explicit MemorySaverResourceView(const int64_t memory_savings_bytes);
 
   MemorySaverResourceView(const MemorySaverResourceView&) = delete;
   MemorySaverResourceView& operator=(const MemorySaverResourceView&) = delete;
